+++ compute_profile.h
@@ -7,6 +7,6 @@ using namespace std;
 
 //--------------- recentering coordinates---------------
 void ro_r(const vector <float> x, const vector <float> y, const vector <float> z,
-        const int nrings, const float max_distance, vector <float> &ro);
+        const int nrings, const float max_distance, vector <double> &ro);
 
 #endif
